requirements.txt
transformers>=4.36.2
torch>=2.2.0
pydantic>=2.5.3
pyyaml>=6.0.1
pytest>=7.4.3
python-dotenv>=1.0.0
